requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
google-cloud-storage==2.14.0
requests==2.31.0
pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
